Add css link to /default.css
The idea here is to use the "default" css for whatever site is
hosting this.

Whether or not a given site has a /default.css is an exercise
left to the reader.  Afaik, there's no general-convention
on what the default css for a site is, but I propose that
default.css is not a bad convention.