Fix how ?today works to avoid repeats
Instead of getting the page of the day using the random number
from the time of the start of the day as a seed, instead,
shuffle the chapters based on a random seed = start of the year,
and pick the nth chapter from (day of year)%(number of chapters).
This avoids unnecessary repeats (though obviously there will be
repeats within a given year, since there are fewer chapters
than days in the year).

It's not efficient, but that doesn't matter.